refactor(cockpit): extract proof evidence artifact parsing
